Refactor code layout
