+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 //import { createStore } from 'redux';
 //import reducer from './redux/reducers';
-import App from './App';
+import App from './components/App';
 import './index.css';
 
 // const store = createStore(
